WIP realized I have to implement another client manager strategy
