+++ examples/feddg_ga_example/server.py
@@ -11,7 +11,7 @@ from examples.models.cnn_model import MnistNetWithBnAndFrozen
 from examples.simple_metric_aggregation import evaluate_metrics_aggregation_fn, fit_metrics_aggregation_fn
 from examples.utils.functions import make_dict_with_epochs_or_steps
 from fl4health.model_bases.apfl_base import ApflModule
-from fl4health.server.base_server import FedDGGAServer
+from fl4health.server.feddg_ga_server import FedDGGAServer
 from fl4health.utils.config import load_config
 from fl4health.utils.functions import get_all_model_parameters
 from fl4health.utils.random import set_all_random_seeds
